Remove attributes in favour of properties
This was neccessary as attributes didn't make sense in
the concept of a "schema" domain, whereas a "property"
did. In addition, it was deemed way to complicated to make
attributes reuseable (while still having logic) across
fields. In the end, it made sense to make properties
behave more like dumb wrappers.